Fixed Warning - in register_serialize_entry.
Ho effettuato un cast tra tipi interi per evitare l'insorgere di un warning.

+++ register.c
@@ -186,7 +186,7 @@ char* register_serialize_entry(const struct entry* E, char* buf, size_t len)
 
     if (buf != NULL)
     {
-        if (len < DATE_LENGHT+2+TYPE_LENGHT+strLen+1)
+        if ((int)len < DATE_LENGHT+2+TYPE_LENGHT+strLen+1)
         {
             /** Se non basta lo spazio per contenere
              * il risultato genera un errore
